move duplicated code to mixin
